Refactor: moved ETA formatting to directions utils
